universe_main/requirements.txt
django==3.2
psycopg2-binary==2.8.6
pillow==8.2
django-rest-framework==0.1.0
django-allauth==0.44.0
django-rest-swagger==2.2.0
redis==3.5.3